Trying to typedast size in malloc

+++ 105-radix_sort.c
@@ -62,8 +62,8 @@ void countingSort(int *array, int size, int place)
 	int max, i, *output, *count;
 
 	max = (array[0] / place) % 10;
-	output = malloc(sizeof(array[0]) * size);
-	count = malloc(sizeof(array[0]) * size);
+	output = malloc(sizeof(array[0]) * (int)size);
+	count = malloc(sizeof(array[0]) * (int)size);
 
 	for (i = 1; i < size; i++)
 	{
